fix(client): correct 5 parity bugs + 2 edge-case value divergences vs Java rev-225
Found by the 2026-05-28 line-by-line Go<->Java(225-clean) parity audit.

- load(): wrap the mapscene/mapfunction sprite loops in per-loop panic recovery
  (matching Java's per-loop try/catch) so a media archive with fewer than 50
  sprites leaves the missing entries nil and continues, instead of aborting boot.
- gameshell: call InputTracking.KeyPressed on every key press (not only sentinel
  keys) so arrow/modifier keys are recorded like Java; charFor returns Java's
  CHAR_UNDEFINED (65535) for non-text keys.
- DrawInterface type-6 and ObjType.GetIcon: truncate sinTable*zoom to int32
  before >>16 so large-zoom interface models render at Java's angle/scale (Go's
  64-bit int otherwise skips the 32-bit overflow/wrap).
- signlink.GetUID: guard a short/corrupt uid.dat (return 0 like Java's caught
  EOFException) instead of panicking; adds a regression test.
- ::clientdrop: always reconnect in the standalone launch mode (Java
  super.frame \!= null is always true there), instead of requiring a 192.168.1.x host.
- chatback amount: parse with ParseInt(base 10, 32-bit) so out-of-range values
  send 0 like Java's Integer.parseInt, instead of a bit-truncated nonzero P4.
- examine descriptions: transcode via io.Latin1ToUTF8 (Java new String(byte[])
  default/Latin-1 charset) so non-ASCII bytes render correctly, not as U+FFFD.

+++ pkg/jagex2/config/objtype/objtype.go
@@ -356,8 +356,12 @@ func GetIcon(arg0, arg2 int) *pix32.Pix32 {
 	pix2d.FillRect(0, 0, 0, 32, 32)
 	pix3d.Init2D()
 	var15 := var4.GetInterfaceModel(1)
-	var16 := (pix3d.SinTable[var4.Xan2D] * var4.Zoom2D) >> 16
-	var17 := (pix3d.CosTable[var4.Xan2D] * var4.Zoom2D) >> 16
+	// Java: `Pix3D.sinTable[xan2d] * zoom2d >> 16` is 32-bit int arithmetic; the
+	// product overflows/wraps at 2^31 (reachable when zoom2d > 32768). int32(...)
+	// reproduces that truncation before the arithmetic >>16, which Go's 64-bit int
+	// would otherwise skip. Same fix as DrawInterface type-6 (client.go).
+	var16 := int(int32(pix3d.SinTable[var4.Xan2D]*var4.Zoom2D)) >> 16
+	var17 := int(int32(pix3d.CosTable[var4.Xan2D]*var4.Zoom2D)) >> 16
 	var15.DrawSimple(0, var4.Yan2D, var4.Zan2D, var4.Xan2D, var4.Xof2D, var16+var15.MaxY/2+var4.Yof2D, var17+var4.Yof2D)
 	for i := 31; i >= 0; i-- {
 		for j := 31; j >= 0; j-- {
